implement initial version of cache max transaction flow calculation

+++ src/core/transactions/transactions/max_flow_calculation/InitiateMaxFlowCalculationTransaction.h
@@ -4,7 +4,6 @@
 #include "../base/BaseTransaction.h"
 #include "../../../interface/commands_interface/commands/max_flow_calculation/InitiateMaxFlowCalculationCommand.h"
 
-#include "../../scheduler/TransactionsScheduler.h"
 #include "../../../trust_lines/manager/TrustLinesManager.h"
 #include "../../../max_flow_calculation/manager/MaxFlowCalculationTrustLineManager.h"
 #include "../../../network/messages/outgoing/max_flow_calculation/InitiateMaxFlowCalculationMessage.h"
@@ -38,7 +37,7 @@ private:
     TrustLineAmount calculateOneNode(
         const NodeUUID& nodeUUID,
         const TrustLineAmount& currentFlow,
-        int level,
+        byte level,
         const NodeUUID& targetUUID,
         const NodeUUID& sourceUUID);
